Error handler on $axios instances
