+++ plugins/europeana/recommendation.js
@@ -1,9 +1,12 @@
-import { apiError, createAxios } from './utils';
+import { apiError, createAxios, keycloakErrorHandler } from './utils';
 
 export const BASE_URL = process.env.EUROPEANA_RECOMMENDATION_API_URL || 'https://api.europeana.eu/recommend';
 
 export default (context = {}) => {
-  const $axios = createAxios({ id: 'recommendation', baseURL: BASE_URL, $axios: context.$axios }, context);
+  const $axios = createAxios(
+    { id: 'recommendation', baseURL: BASE_URL, $axios: context.$axios, errorHandler: keycloakErrorHandler },
+    context
+  );
 
   return {
     $axios,
